Changed the Folder Structure and Updated the code on index, GridIcon and ListIcon
