+++ client/src/shared/icons/ListIcon.tsx
@@ -4,18 +4,20 @@ import type { FC } from 'react';
 export interface ListIconProps {
   height?: number,
   width?: number,
-  children?: string
+  classname?: string
 }
 
-const ListIcon: FC<ListIconProps> = ({ height, width, children }: ListIconProps) => {
+const ListIcon: FC<ListIconProps> = ({ height, width, classname }: ListIconProps) => {
   return (
     <svg
+      height={height}
+      width={width}
       xmlns="http://www.w3.org/2000/svg"
       fill="none"
       viewBox="0 0 24 24"
       strokeWidth={1.5}
       stroke="black"
-      className={`w-${width} h-${height} ${children}`}
+      className={classname}
     >
       <path
         strokeLinecap="round"
@@ -27,9 +29,9 @@ const ListIcon: FC<ListIconProps> = ({ height, width, children }: ListIconProps)
 };
 
 ListIcon.defaultProps = {
-  height: 6,
-  width: 6,
-  children: ''
+  height: 25,
+  width: 25,
+  classname: ''
 };
 
 export default ListIcon;
